chore: separate telemetry from onboarding
