+++ apps/trading/components/telemetry/index.ts
@@ -1,0 +1,1 @@
+export { Telemetry } from './telemetry';
