+++ apps/trading/components/telemetry/telemetry-approval.spec.tsx
