+++ apps/trading/components/telemetry/telemetry-approval.tsx
